Now possible to test and refine QUST-side dialogue management
Fixed loading and saving of DIAL/TIFC.

The Log Window now clears the selected log item text when you clear all log items.

Dialogue models: implemented indexing branchless topics in the datastore; implemented filtering branchless topics in the model.

Various fixes to the dialogue UI. It's now possible to browse a quest's dialogue! Still need to implement form editing dialogs for DLBR, DIAL, and INFO, as well as a plethora of fixes to the browser on QUST.

+++ native/src/DovahKit/ui/form_windows/quest/QuestAllDialogueDatastore.h
@@ -38,7 +38,7 @@ class QuestAllDialogueDatastore : public QObject {
             bool     has_end_fragment    = false;
             bool     has_own_prompt      = false;
             bool     uses_shared_info    = false;
-            uint16_t hours_until_reset   = 0;
+            float    hours_until_reset   = 0;
             uint32_t flags = 0;
          } cached;
